# docs/swagger.yaml
basePath: /
definitions:
  controller.JamKerja:
    properties:
      durasi:
        example: 8
        type: integer
      gmt:
        example: 7
        type: integer
      hari:
        example:
        - Senin
        - Selasa
        - Rabu
        - Kamis
        - Jumat
        - Sabtu
        - Minggu
        items:
          type: string
        type: array
      jam_keluar:
        example: "16:00"
        type: string
      jam_masuk:
        example: "08:00"
        type: string
      piket_tim:
        example: Piket Z
        type: string
      shift:
        example: 2
        type: integer
    type: object
  controller.Karyawan:
    properties:
      _id:
        example: "123456789"
        type: string
      hari_kerja:
        example:
        - Senin
        - Selasa
        - Rabu
        - Kamis
        - Jumat
        - Sabtu
        - Minggu
        items:
          type: string
        type: array
      jabatan:
        example: Anonymous
        type: string
      jam_kerja:
        items:
          $ref: '#/definitions/controller.JamKerja'
        type: array
      nama:
        example: Tes Swagger
        type: string
      phone_number:
        example: "08123456789"
        type: string
    type: object
  controller.Presensi:
    properties:
      _id:
        example: "123456789"
        type: string
      biodata:
        $ref: '#/definitions/controller.Karyawan'
      checkin:
        example: Masuk
        type: string
      datetime:
        example: "2024-09-01T00:00:00Z"
        format: date-time
        type: string
      latitude:
        example: 123.12
        type: number
      location:
        example: Bandung
        type: string
      longitude:
        example: 123.11
        type: number
      phone_number:
        example: "08123456789"
        type: string
    type: object
  controller.ReqKaryawan:
    properties:
      hari_kerja:
        example:
        - Senin
        - Selasa
        - Rabu
        - Kamis
        - Jumat
        - Sabtu
        - Minggu
        items:
          type: string
        type: array
      jabatan:
        example: Anonymous
        type: string
      jam_kerja:
        items:
          $ref: '#/definitions/controller.JamKerja'
        type: array
      nama:
        example: Tes Swagger
        type: string
      phone_number:
        example: "08123456789"
        type: string
    type: object
  controller.ReqPresensi:
    properties:
      biodata:
        $ref: '#/definitions/controller.ReqKaryawan'
      checkin:
        example: Masuk
        type: string
      latitude:
        example: 123.12
        type: number
      location:
        example: Bandung
        type: string
      longitude:
        example: 123.11
        type: number
      phone_number:
        example: "08123456789"
        type: string
    type: object
host: ws-deviwlndr-59b3a0157dd7.herokuapp.com
info:
  contact:
    email: 714220054@std.ulbi.ac.id
    name: API Support
    url: https://github.com/deviwlndr
  description: This is a sample swagger for Fiber
  title: TES SWAGGER ULBI
  version: "1.0"
paths:
  /delete/{id}:
    delete:
      consumes:
      - application/json
      description: Hapus data presensi.
      parameters:
      - description: Masukan ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Delete data presensi.
      tags:
      - Presensi
  /insert:
    post:
      consumes:
      - application/json
      description: Input data presensi.
      parameters:
      - description: Payload Body [RAW]
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controller.ReqPresensi'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.Presensi'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Insert data presensi.
      tags:
      - Presensi
  /presensi:
    get:
      consumes:
      - application/json
      description: Mengambil semua data presensi.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.Presensi'
      summary: Get All Data Presensi.
      tags:
      - Presensi
  /presensi/{id}:
    get:
      consumes:
      - application/json
      description: Ambil per ID data presensi.
      parameters:
      - description: Masukan ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.Presensi'
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Get By ID Data Presensi.
      tags:
      - Presensi
  /update/{id}:
    put:
      consumes:
      - application/json
      description: Ubah data presensi.
      parameters:
      - description: Masukan ID
        in: path
        name: id
        required: true
        type: string
      - description: Payload Body [RAW]
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controller.ReqPresensi'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.Presensi'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Update data presensi.
      tags:
      - Presensi
schemes:
- https
- http
swagger: "2.0"